+++ aws/resource_aws_workspaces_directory_test.go
@@ -7,52 +7,66 @@ import (
 	"testing"
 
 	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/iam"
 	"github.com/aws/aws-sdk-go/service/workspaces"
-	multierror "github.com/hashicorp/go-multierror"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
+	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/workspaces/finder"
+	"github.com/terraform-providers/terraform-provider-aws/aws/internal/tfresource"
 )
 
 func init() {
 	resource.AddTestSweepers("aws_workspaces_directory", &resource.Sweeper{
 		Name:         "aws_workspaces_directory",
 		F:            testSweepWorkspacesDirectories,
-		Dependencies: []string{"aws_workspaces_workspace"},
+		Dependencies: []string{"aws_workspaces_workspace", "aws_workspaces_ip_group"},
 	})
 }
 
 func testSweepWorkspacesDirectories(region string) error {
 	client, err := sharedClientForRegion(region)
 	if err != nil {
-		return fmt.Errorf("error getting client: %w", err)
+		return fmt.Errorf("error getting client: %s", err)
 	}
 	conn := client.(*AWSClient).workspacesconn
-
-	var errors error
 	input := &workspaces.DescribeWorkspaceDirectoriesInput{}
-	err = conn.DescribeWorkspaceDirectoriesPages(input, func(resp *workspaces.DescribeWorkspaceDirectoriesOutput, _ bool) bool {
-		for _, directory := range resp.Directories {
-			err := workspacesDirectoryDelete(aws.StringValue(directory.DirectoryId), conn)
-			if err != nil {
-				errors = multierror.Append(errors, err)
-			}
+	sweepResources := make([]*testSweepResource, 0)
+
+	err = conn.DescribeWorkspaceDirectoriesPages(input, func(page *workspaces.DescribeWorkspaceDirectoriesOutput, lastPage bool) bool {
+		if page == nil {
+			return !lastPage
 		}
-		return true
+
+		for _, directory := range page.Directories {
+			r := resourceAwsWorkspacesDirectory()
+			d := r.Data(nil)
+			d.SetId(aws.StringValue(directory.DirectoryId))
+
+			sweepResources = append(sweepResources, NewTestSweepResource(r, d, client))
+		}
+
+		return !lastPage
 	})
+
 	if testSweepSkipSweepError(err) {
-		log.Printf("[WARN] Skipping Workspace Directory sweep for %s: %s", region, err)
-		return errors // In case we have completed some pages, but had errors
-	}
-	if err != nil {
-		errors = multierror.Append(errors, fmt.Errorf("error listing WorkSpaces Directories: %w", err))
+		log.Printf("[WARN] Skipping WorkSpaces Directory sweep for %s: %s", region, err)
+		return nil
 	}
 
-	return errors
+	if err != nil {
+		return fmt.Errorf("error listing WorkSpaces Directories (%s): %w", region, err)
+	}
+
+	err = testSweepResourceOrchestrator(sweepResources)
+
+	if err != nil {
+		return fmt.Errorf("error sweeping WorkSpaces Directories (%s): %w", region, err)
+	}
+
+	return nil
 }
 
-func TestAccAwsWorkspacesDirectory_basic(t *testing.T) {
+func testAccAwsWorkspacesDirectory_basic(t *testing.T) {
 	var v workspaces.WorkspaceDirectory
 	rName := acctest.RandString(8)
 
@@ -62,7 +76,7 @@ func TestAccAwsWorkspacesDirectory_basic(t *testing.T) {
 
 	domain := testAccRandomDomainName()
 
-	resource.ParallelTest(t, resource.TestCase{
+	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
 			testAccPreCheckWorkspacesDirectory(t)
@@ -98,6 +112,7 @@ func TestAccAwsWorkspacesDirectory_basic(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceName, "workspace_access_properties.0.device_type_android", "ALLOW"),
 					resource.TestCheckResourceAttr(resourceName, "workspace_access_properties.0.device_type_chromeos", "ALLOW"),
 					resource.TestCheckResourceAttr(resourceName, "workspace_access_properties.0.device_type_ios", "ALLOW"),
+					resource.TestCheckResourceAttr(resourceName, "workspace_access_properties.0.device_type_linux", "DENY"),
 					resource.TestCheckResourceAttr(resourceName, "workspace_access_properties.0.device_type_osx", "ALLOW"),
 					resource.TestCheckResourceAttr(resourceName, "workspace_access_properties.0.device_type_web", "DENY"),
 					resource.TestCheckResourceAttr(resourceName, "workspace_access_properties.0.device_type_windows", "ALLOW"),
@@ -120,7 +135,7 @@ func TestAccAwsWorkspacesDirectory_basic(t *testing.T) {
 	})
 }
 
-func TestAccAwsWorkspacesDirectory_disappears(t *testing.T) {
+func testAccAwsWorkspacesDirectory_disappears(t *testing.T) {
 	var v workspaces.WorkspaceDirectory
 	rName := acctest.RandString(8)
 
@@ -128,7 +143,7 @@ func TestAccAwsWorkspacesDirectory_disappears(t *testing.T) {
 
 	domain := testAccRandomDomainName()
 
-	resource.ParallelTest(t, resource.TestCase{
+	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
 			testAccPreCheckWorkspacesDirectory(t)
@@ -143,7 +158,7 @@ func TestAccAwsWorkspacesDirectory_disappears(t *testing.T) {
 				Config: testAccWorkspacesDirectoryConfig(rName, domain),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					testAccCheckAwsWorkspacesDirectoryExists(resourceName, &v),
-					testAccCheckAwsWorkspacesDirectoryDisappears(&v),
+					testAccCheckResourceDisappears(testAccProvider, resourceAwsWorkspacesDirectory(), resourceName),
 				),
 				ExpectNonEmptyPlan: true,
 			},
@@ -151,7 +166,7 @@ func TestAccAwsWorkspacesDirectory_disappears(t *testing.T) {
 	})
 }
 
-func TestAccAwsWorkspacesDirectory_subnetIds(t *testing.T) {
+func testAccAwsWorkspacesDirectory_subnetIds(t *testing.T) {
 	var v workspaces.WorkspaceDirectory
 	rName := acctest.RandString(8)
 
@@ -159,7 +174,7 @@ func TestAccAwsWorkspacesDirectory_subnetIds(t *testing.T) {
 
 	domain := testAccRandomDomainName()
 
-	resource.ParallelTest(t, resource.TestCase{
+	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
 			testAccPreCheckWorkspacesDirectory(t)
@@ -186,7 +201,7 @@ func TestAccAwsWorkspacesDirectory_subnetIds(t *testing.T) {
 	})
 }
 
-func TestAccAwsWorkspacesDirectory_tags(t *testing.T) {
+func testAccAwsWorkspacesDirectory_tags(t *testing.T) {
 	var v workspaces.WorkspaceDirectory
 	rName := acctest.RandString(8)
 
@@ -194,7 +209,7 @@ func TestAccAwsWorkspacesDirectory_tags(t *testing.T) {
 
 	domain := testAccRandomDomainName()
 
-	resource.ParallelTest(t, resource.TestCase{
+	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
 			testAccPreCheckWorkspacesDirectory(t)
@@ -239,7 +254,7 @@ func TestAccAwsWorkspacesDirectory_tags(t *testing.T) {
 	})
 }
 
-func TestAccAwsWorkspacesDirectory_selfServicePermissions(t *testing.T) {
+func testAccAwsWorkspacesDirectory_selfServicePermissions(t *testing.T) {
 	var v workspaces.WorkspaceDirectory
 	rName := acctest.RandString(8)
 
@@ -247,7 +262,7 @@ func TestAccAwsWorkspacesDirectory_selfServicePermissions(t *testing.T) {
 
 	domain := testAccRandomDomainName()
 
-	resource.ParallelTest(t, resource.TestCase{
+	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
 			testAccPreCheckWorkspacesDirectory(t)
@@ -274,7 +289,7 @@ func TestAccAwsWorkspacesDirectory_selfServicePermissions(t *testing.T) {
 	})
 }
 
-func TestAccAwsWorkspacesDirectory_workspaceAccessProperties(t *testing.T) {
+func testAccAwsWorkspacesDirectory_workspaceAccessProperties(t *testing.T) {
 	var v workspaces.WorkspaceDirectory
 	rName := acctest.RandString(8)
 
@@ -282,7 +297,7 @@ func TestAccAwsWorkspacesDirectory_workspaceAccessProperties(t *testing.T) {
 
 	domain := testAccRandomDomainName()
 
-	resource.ParallelTest(t, resource.TestCase{
+	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
 			testAccPreCheckWorkspacesDirectory(t)
@@ -301,6 +316,7 @@ func TestAccAwsWorkspacesDirectory_workspaceAccessProperties(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceName, "workspace_access_properties.0.device_type_android", "ALLOW"),
 					resource.TestCheckResourceAttr(resourceName, "workspace_access_properties.0.device_type_chromeos", "ALLOW"),
 					resource.TestCheckResourceAttr(resourceName, "workspace_access_properties.0.device_type_ios", "ALLOW"),
+					resource.TestCheckResourceAttr(resourceName, "workspace_access_properties.0.device_type_linux", "DENY"),
 					resource.TestCheckResourceAttr(resourceName, "workspace_access_properties.0.device_type_osx", "ALLOW"),
 					resource.TestCheckResourceAttr(resourceName, "workspace_access_properties.0.device_type_web", "DENY"),
 					resource.TestCheckResourceAttr(resourceName, "workspace_access_properties.0.device_type_windows", "DENY"),
@@ -311,7 +327,7 @@ func TestAccAwsWorkspacesDirectory_workspaceAccessProperties(t *testing.T) {
 	})
 }
 
-func TestAccAwsWorkspacesDirectory_workspaceCreationProperties(t *testing.T) {
+func testAccAwsWorkspacesDirectory_workspaceCreationProperties(t *testing.T) {
 	var v workspaces.WorkspaceDirectory
 	rName := acctest.RandString(8)
 
@@ -320,7 +336,7 @@ func TestAccAwsWorkspacesDirectory_workspaceCreationProperties(t *testing.T) {
 
 	domain := testAccRandomDomainName()
 
-	resource.ParallelTest(t, resource.TestCase{
+	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
 			testAccPreCheckWorkspacesDirectory(t)
@@ -347,7 +363,7 @@ func TestAccAwsWorkspacesDirectory_workspaceCreationProperties(t *testing.T) {
 	})
 }
 
-func TestAccAwsWorkspacesDirectory_workspaceCreationProperties_customSecurityGroupId_defaultOu(t *testing.T) {
+func testAccAwsWorkspacesDirectory_workspaceCreationProperties_customSecurityGroupId_defaultOu(t *testing.T) {
 	var v workspaces.WorkspaceDirectory
 	rName := acctest.RandString(8)
 
@@ -356,7 +372,7 @@ func TestAccAwsWorkspacesDirectory_workspaceCreationProperties_customSecurityGro
 
 	domain := testAccRandomDomainName()
 
-	resource.ParallelTest(t, resource.TestCase{
+	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
 			testAccPreCheckWorkspacesDirectory(t)
@@ -399,7 +415,7 @@ func TestAccAwsWorkspacesDirectory_workspaceCreationProperties_customSecurityGro
 	})
 }
 
-func TestAccAwsWorkspacesDirectory_ipGroupIds(t *testing.T) {
+func testAccAwsWorkspacesDirectory_ipGroupIds(t *testing.T) {
 	var v workspaces.WorkspaceDirectory
 	rName := acctest.RandString(8)
 
@@ -407,7 +423,7 @@ func TestAccAwsWorkspacesDirectory_ipGroupIds(t *testing.T) {
 
 	domain := testAccRandomDomainName()
 
-	resource.ParallelTest(t, resource.TestCase{
+	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t); testAccPreCheckHasIAMRole(t, "workspaces_DefaultRole") },
 		ErrorCheck:   testAccErrorCheck(t, workspaces.EndpointsID),
 		Providers:    testAccProviders,
@@ -539,6 +555,7 @@ func TestExpandWorkspaceAccessProperties(t *testing.T) {
 					"device_type_android":    "ALLOW",
 					"device_type_chromeos":   "ALLOW",
 					"device_type_ios":        "ALLOW",
+					"device_type_linux":      "DENY",
 					"device_type_osx":        "ALLOW",
 					"device_type_web":        "DENY",
 					"device_type_windows":    "DENY",
@@ -549,6 +566,7 @@ func TestExpandWorkspaceAccessProperties(t *testing.T) {
 				DeviceTypeAndroid:    aws.String("ALLOW"),
 				DeviceTypeChromeOs:   aws.String("ALLOW"),
 				DeviceTypeIos:        aws.String("ALLOW"),
+				DeviceTypeLinux:      aws.String("DENY"),
 				DeviceTypeOsx:        aws.String("ALLOW"),
 				DeviceTypeWeb:        aws.String("DENY"),
 				DeviceTypeWindows:    aws.String("DENY"),
@@ -581,6 +599,7 @@ func TestFlattenWorkspaceAccessProperties(t *testing.T) {
 				DeviceTypeAndroid:    aws.String("ALLOW"),
 				DeviceTypeChromeOs:   aws.String("ALLOW"),
 				DeviceTypeIos:        aws.String("ALLOW"),
+				DeviceTypeLinux:      aws.String("DENY"),
 				DeviceTypeOsx:        aws.String("ALLOW"),
 				DeviceTypeWeb:        aws.String("DENY"),
 				DeviceTypeWindows:    aws.String("DENY"),
@@ -591,6 +610,7 @@ func TestFlattenWorkspaceAccessProperties(t *testing.T) {
 					"device_type_android":    "ALLOW",
 					"device_type_chromeos":   "ALLOW",
 					"device_type_ios":        "ALLOW",
+					"device_type_linux":      "DENY",
 					"device_type_osx":        "ALLOW",
 					"device_type_web":        "DENY",
 					"device_type_windows":    "DENY",
@@ -705,25 +725,6 @@ func TestFlattenWorkspaceCreationProperties(t *testing.T) {
 	}
 }
 
-func testAccPreCheckHasIAMRole(t *testing.T, roleName string) {
-	conn := testAccProvider.Meta().(*AWSClient).iamconn
-
-	input := &iam.GetRoleInput{
-		RoleName: aws.String(roleName),
-	}
-	_, err := conn.GetRole(input)
-
-	if isAWSErr(err, iam.ErrCodeNoSuchEntityException, "") {
-		t.Skipf("skipping acceptance test: required IAM role \"%s\" is not present", roleName)
-	}
-	if testAccPreCheckSkipError(err) {
-		t.Skipf("skipping acceptance test: %s", err)
-	}
-	if err != nil {
-		t.Fatalf("unexpected PreCheck error: %s", err)
-	}
-}
-
 func testAccCheckAwsWorkspacesDirectoryDestroy(s *terraform.State) error {
 	conn := testAccProvider.Meta().(*AWSClient).workspacesconn
 
@@ -732,57 +733,44 @@ func testAccCheckAwsWorkspacesDirectoryDestroy(s *terraform.State) error {
 			continue
 		}
 
-		resp, err := conn.DescribeWorkspaceDirectories(&workspaces.DescribeWorkspaceDirectoriesInput{
-			DirectoryIds: []*string{aws.String(rs.Primary.ID)},
-		})
+		_, err := finder.DirectoryByID(conn, rs.Primary.ID)
+
+		if tfresource.NotFound(err) {
+			continue
+		}
+
 		if err != nil {
 			return err
 		}
 
-		if len(resp.Directories) == 0 {
-			return nil
-		}
-
-		dir := resp.Directories[0]
-		if *dir.State != workspaces.WorkspaceDirectoryStateDeregistering && *dir.State != workspaces.WorkspaceDirectoryStateDeregistered {
-			return fmt.Errorf("directory %q was not deregistered", rs.Primary.ID)
-		}
+		return fmt.Errorf("WorkSpaces Directory %s still exists", rs.Primary.ID)
 	}
 
 	return nil
-}
-
-func testAccCheckAwsWorkspacesDirectoryDisappears(v *workspaces.WorkspaceDirectory) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		return workspacesDirectoryDelete(aws.StringValue(v.DirectoryId), testAccProvider.Meta().(*AWSClient).workspacesconn)
-	}
 }
 
 func testAccCheckAwsWorkspacesDirectoryExists(n string, v *workspaces.WorkspaceDirectory) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		rs, ok := s.RootModule().Resources[n]
 		if !ok {
-			return fmt.Errorf("workspaces directory resource is not found: %q", n)
+			return fmt.Errorf("Not found: %s", n)
 		}
 
 		if rs.Primary.ID == "" {
-			return fmt.Errorf("workspaces directory resource ID is not set")
+			return fmt.Errorf("No WorkSpaces Directory ID is set")
 		}
 
 		conn := testAccProvider.Meta().(*AWSClient).workspacesconn
-		resp, err := conn.DescribeWorkspaceDirectories(&workspaces.DescribeWorkspaceDirectoriesInput{
-			DirectoryIds: []*string{aws.String(rs.Primary.ID)},
-		})
+
+		output, err := finder.DirectoryByID(conn, rs.Primary.ID)
+
 		if err != nil {
 			return err
 		}
 
-		if *resp.Directories[0].DirectoryId == rs.Primary.ID {
-			*v = *resp.Directories[0]
-			return nil
-		}
+		*v = *output
 
-		return fmt.Errorf("workspaces directory %q is not found", rs.Primary.ID)
+		return nil
 	}
 }
 
@@ -957,6 +945,7 @@ resource "aws_workspaces_directory" "main" {
     device_type_android    = "ALLOW"
     device_type_chromeos   = "ALLOW"
     device_type_ios        = "ALLOW"
+    device_type_linux      = "DENY"
     device_type_osx        = "ALLOW"
     device_type_web        = "DENY"
     device_type_windows    = "DENY"